requirements.txt
requests~=2.25
PyYAML~=5.4
